Move line to a sub directory
